% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSMOOTH.R
\name{GMSMOOTH}
\alias{GMSMOOTH}
\title{Nonparametric kernel estimation}
\usage{
GMSMOOTH(y, x, h)
}
\arguments{
\item{y}{True distributions}
\item{x}{ A random seq}

\item{h}{the bandwidth}
}
\value{
GM kernel estimation \code{n}
}
\description{
two methods of Nonparametric kernel estimation
}
\examples{
\dontrun{
x <- seq(-1, 1, length = 40)
y <- 5 * x * cos(5 * pi * x)
h <- 0.055
GMsmooth.val <- GMSMOOTH(y, x, h)
}
}
