% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWSMOOTH.R
\name{NWSMOOTH}
\alias{NWSMOOTH}
\title{Nonparametric kernel estimation}
\usage{
NWSMOOTH(h, y, x)
}
\arguments{
\item{h}{The bandwidth}

\item{y}{True distributions}
\item{x}{A random seq}
}
\value{
NW kernel estimation \code{n}
}
\description{
two methods of Nonparametric kernel estimation
}
\examples{
\dontrun{
x <- seq(-1, 1, length = 40)
y <- 5 * x * cos(5 * pi * x)
h <- 0.055
NWsmooth.val <- NWSMOOTH(h, y, x)
}
}
